requirements.txt
Pillow==9.3.0
Pillow==10.2.0
rembg==2.0.55
streamlit==1.31.0
